refactor: move responsive button to ui directory
